Added login and registration pages, as well as pages for recommended books, my library and a page for reading a book.  On the "recommended books page" realized filter, transition to the page with the library, as well as the list of recommended books. Realized modal window for adding a book to your library.  On the page "my library" implemented a form to add a book and a list of recommended books with a transition to the "recommended page". Also realized rendering of books that are added to the library.  When you click on the added book opens a modal window to go to the page with reading this book. On this page is realized the function of starting and stopping the reading, as well as statistics on the dates of reading and how many pages of the book read.
Added login and registration pages, as well as pages for recommended books, my library and a page for reading a book.
On the "recommended books page" realized filter, transition to the page with the library, as well as the list of recommended books. Realized modal window for adding a book to your library.
On the page "my library" implemented a form to add a book and a list of recommended books with a transition to the "recommended page". Also realized rendering of books that are added to the library.
When you click on the added book opens a modal window to go to the page with reading this book. On this page is realized the function of starting and stopping the reading, as well as statistics on the dates of reading and how many pages of the book read.